Make DiscFilterResponse concrete and remove package-private sub-class
Make DiscFilterResponse opereate directly on a jdisc-core Response
instance. Reduce use of jdisc-core types from its public API.
