+++ container-core/abi-spec.json
@@ -2247,8 +2247,7 @@
     "superClass": "java.lang.Object",
     "interfaces": [],
     "attributes": [
-      "public",
-      "abstract"
+      "public"
     ],
     "methods": [
       "public void <init>(com.yahoo.jdisc.http.HttpResponse)",
@@ -2258,16 +2257,16 @@
       "public void removeAttribute(java.lang.String)",
       "public com.yahoo.jdisc.HeaderFields getUntreatedHeaders()",
       "public java.util.List getUntreatedCookies()",
-      "public abstract void setHeader(java.lang.String, java.lang.String)",
-      "public abstract void removeHeaders(java.lang.String)",
-      "public abstract void setHeaders(java.lang.String, java.lang.String)",
-      "public abstract void setHeaders(java.lang.String, java.util.List)",
-      "public abstract void addHeader(java.lang.String, java.lang.String)",
-      "public abstract java.lang.String getHeader(java.lang.String)",
+      "public void setHeader(java.lang.String, java.lang.String)",
+      "public void removeHeaders(java.lang.String)",
+      "public void setHeaders(java.lang.String, java.lang.String)",
+      "public void setHeaders(java.lang.String, java.util.List)",
+      "public void addHeader(java.lang.String, java.lang.String)",
+      "public java.lang.String getHeader(java.lang.String)",
       "public java.util.List getCookies()",
-      "public abstract void setCookies(java.util.List)",
+      "public void setCookies(java.util.List)",
       "public int getStatus()",
-      "public abstract void setStatus(int)",
+      "public void setStatus(int)",
       "public com.yahoo.jdisc.http.HttpResponse getParentResponse()",
       "public void addCookie(com.yahoo.jdisc.http.filter.JDiscCookieWrapper)",
       "public void sendError(int)",
